@@ -4,80 +4,86 @@
   "description": "Boilerplate for react, redux with server-side rendering and support for progressive web application",
   "main": "index.js",
   "engines": {
-    "node": ">= 7.9.0"
+    "node": ">= 8.4.0"
   },
   "engineStrict": true,
   "devDependencies": {
     "assets-webpack-plugin": "^3.5.1",
-    "autoprefixer": "^7.1.2",
-    "babel-core": "^6.25.0",
-    "babel-eslint": "^7.2.3",
-    "babel-loader": "^7.1.0",
+    "autoprefixer": "^7.1.4",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.0.0",
+    "babel-loader": "^7.1.2",
     "babel-plugin-css-modules-transform": "^1.2.7",
     "babel-plugin-lodash": "^3.2.11",
-    "babel-plugin-react-transform": "^2.0.2",
+    "babel-plugin-react-transform": "^3.0.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-object-rest-spread": "^6.23.0",
-    "babel-polyfill": "^6.23.0",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.6.0",
     "babel-preset-react": "^6.24.1",
-    "babel-register": "^6.24.1",
-    "chai": "^4.1.0",
+    "babel-register": "^6.26.0",
+    "chai": "^4.1.2",
     "chokidar": "^1.7.0",
     "classnames": "^2.2.5",
     "clean-webpack-plugin": "^0.1.16",
     "compression": "^1.7.0",
     "cookie-parser": "^1.4.3",
     "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.28.4",
-    "eslint": "^4.3.0",
+    "cross-env": "^5.0.5",
+    "css-loader": "^0.28.7",
+    "eslint": "^4.6.1",
     "eslint-loader": "^1.9.0",
-    "eslint-plugin-react": "^7.1.0",
-    "express": "^4.15.3",
+    "eslint-plugin-react": "^7.3.0",
+    "express": "^4.15.4",
     "extract-text-webpack-plugin": "^3.0.0",
     "file-loader": "^0.11.2",
     "glob": "^7.1.2",
-    "history": "^4.6.3",
+    "history": "^4.7.2",
     "img-loader": "^2.0.0",
+    "isbot": "^2.1.0",
     "lodash": "^4.17.4",
-    "mocha": "^3.4.2",
+    "memory-cache": "^0.2.0",
+    "mocha": "^3.5.0",
     "moment": "^2.18.1",
+    "multi-loader": "^0.1.0",
     "node-sass": "^4.5.3",
-    "nodemon": "^1.11.0",
+    "nodemon": "^1.12.0",
     "postcss-loader": "^2.0.6",
     "postcss-scss": "^1.0.2",
     "postcss-smart-import": "^0.7.5",
     "precss": "^2.0.0",
     "prop-types": "^15.5.10",
-    "react": "next",
-    "react-dom": "next",
-    "react-hot-loader": "next",
-    "react-redux": "^5.0.5",
-    "react-router": "^4.1.2",
-    "react-router-dom": "^4.1.2",
-    "react-router-redux": "next",
+    "react": "^16.0.0-rc.3",
+    "react-dom": "^16.0.0-rc.2",
+    "react-hot-loader": "^3.0.0-beta.7",
+    "react-redux": "^5.0.6",
+    "react-router": "^4.2.0",
+    "react-router-dom": "^4.2.2",
+    "react-router-redux": "^5.0.0-alpha.6",
     "react-test-renderer": "^15.6.1",
     "react-transform-hmr": "^1.0.4",
     "redux": "^3.7.2",
     "sass-loader": "^6.0.6",
     "serve-favicon": "^2.4.3",
+    "srcset-loader": "^2.1.0",
     "strict-uri-encode": "^2.0.0",
     "style-loader": "^0.18.2",
-    "uglify-js": "^3.0.26",
+    "uglify-js": "^3.0.28",
     "uglifyjs-webpack-plugin": "^0.4.6",
     "universal-fetch": "^1.0.0",
     "url-loader": "^0.5.9",
-    "webp-loader": "^0.2.0",
-    "webpack": "^3.4.1",
-    "webpack-dev-middleware": "^1.11.0",
-    "webpack-dev-server": "^2.6.1",
-    "webpack-hot-middleware": "^2.18.2"
+    "webp-loader": "^0.2.1",
+    "webpack": "^3.5.6",
+    "webpack-delete-after-emit": "^1.0.5",
+    "webpack-dev-middleware": "^1.12.0",
+    "webpack-dev-server": "^2.7.1",
+    "webpack-hot-middleware": "^2.19.1"
   },
   "scripts": {
-    "start": "nodemon --watch src/server.js --watch webpack/dev.babel.js --watch .bin/ ./bin/babelize.js ./dev.server.js",
-    "build": "NODE_ENV=production webpack -p --config ./webpack/prod.client.babel.js && NODE_ENV=production webpack -p --config ./webpack/prod.server.babel.js",
-    "test": "mocha \"test/**/*.test.js\" --require bin/babelize.js --ignore-babelize-require"
+    "start": "cross-env NODE_ENV=development nodemon --watch src/server.js --watch webpack/dev.babel.js --watch .bin/ ./bin/babelize.js ./dev.server.js",
+    "build": "cross-env NODE_ENV=production webpack -p --config ./webpack/prod.client.babel.js && cross-env NODE_ENV=production webpack -p --verbose --config ./webpack/prod.server.babel.js",
+    "test": "cross-env NODE_ENV=test mocha \"test/**/*.test.js\" --require bin/babelize.js --ignore-babelize-require"
   },
   "keywords": [
     "es7",
