@@ -2,7 +2,6 @@
  * @description This file is for registering babel register
  * thus including server file
  */
-/* eslint-disable no-console */
 const fs = require("fs");
 const path = require("path");
 
@@ -12,7 +11,9 @@ let config;
 try {
   config = JSON.parse(babelrc);
 } catch (err) {
+  // eslint-disable-next-line
   console.error("==> ERROR: Error parsing your .babelrc.");
+  // eslint-disable-next-line
   console.error(err);
 }
 
@@ -28,23 +29,11 @@ require("babel-register")(config);
 require("babel-polyfill");
 
 
+const settings = require("../settings");
 const directories = require("../directories");
-
-/**
- * try to include src to path, but with last priority
- */
-try {
-  const extraIncludePaths = [ directories.srcDir ];
-  process.env.NODE_PATH = `${process.env.NODE_PATH}:${extraIncludePaths.join(":")}`;
-  require("module").Module._initPaths();
-} catch (ex) {
-  // Do nothing
-}
-
 const _  = require("lodash");
-const appConfig = require(`${directories.srcDir}/config`);
 
-const allowedImageExtensions = _.get(appConfig, "config.images.allowedExtensions", [
+const allowedImageExtensions = _.get(settings, "images.allowedExtensions", [
   ".jpeg",
   ".jpg",
   ".png",
@@ -55,10 +44,8 @@ const allowedImageExtensions = _.get(appConfig, "config.images.allowedExtensions
 
 _.each(allowedImageExtensions, ext => {
   require.extensions[ext] = function (module, filename) {
-    let name = filename.split("/").pop();
-    name = name.replace(ext, "");
-    const customExt = ext !== ".svg" ? ".webp" : ext;
-    module.exports = `${directories.buildPublicPath}images/${name}${customExt}`;
+    const name = filename.split("/").pop().replace(ext, "");
+    module.exports = `${directories.buildPublicPath}images/${name}${ext}`;
   };
 });
 
@@ -69,9 +56,11 @@ if (process.argv.indexOf("--ignore-babelize-require") === -1) {
     if (fs.existsSync(resolvedFile)) {
       require(resolvedFile);
     } else {
+      // eslint-disable-next-line
       console.log(`Cannot resolve "${relativePathToFile}"`);
+      
+      // eslint-disable-next-line
       console.log(`Make sure the path os relative to ${path.resolve(__dirname)}`);
     }
   }
 }
-/* eslint-enable */
