@@ -7,7 +7,7 @@ import path from "path";
  *    images  --> server images
  *    public --> Public folder that can be pushed to cdn
  *      build --> Compiled files
- *    server.js --> Single compiled server file
+ *    middleware.js --> Single compiled server file
  *  node_modules
  *  src --> Main sources
  *    app
@@ -22,8 +22,8 @@ import path from "path";
  *    utils --> Util functions
  *    client.js --> Client entry js
  *    routes.js --> Consolidated routes
- *    server.js --> Server instance
- *    start-server.js --> Simple start server
+ *    middleware.js --> Server instance
+ *    start-middleware.js --> Simple start server
  */
 
 /**
@@ -31,7 +31,7 @@ import path from "path";
  * application building and in src folder
  * @type {string}
  */
-const publicDirName = "public";
+export const publicDirName = "public";
 
 /**
  * @description Distribution directory name where all the code will
@@ -62,11 +62,16 @@ const buildDirName = "build";
 const srcDirName = "src";
 
 /**
+ * @description the source of all the pages that needs code splitting
+ * @type {string}
+ */
+const pagesDirName = "pages";
+/**
  * @description buildPublicPath is the path that would be used by dev server
  * and the files will be dropped in the path relative to distribution folder
  * @type {string}
  */
-export const buildPublicPath = `/${publicDirName}/${buildDirName}/`;
+export const buildPublicPath = "/" + publicDirName + "/" + buildDirName + "/";
 
 // Directory structure
 // Root dir is the project root
@@ -86,3 +91,5 @@ export const srcPublicDir = path.resolve(path.join(srcDir, publicDirName));
 export const distPublicDir = path.join(distDir, publicDirName);
 
 export const buildDir = path.join(distPublicDir, buildDirName);
+
+export const pagesDir = path.join(srcDir, pagesDirName);
