@@ -1,5 +1,5 @@
 import Api from "./api";
-import storage from "core/libs/storage/storage.server";
+import storage from "../storage";
 
 const apiInstance = new Api({storage});
 export default apiInstance;
