@@ -4,6 +4,12 @@
 import webpack from "webpack";
 
 /**
+ * @description Enable multiple loaders to be executed parallel for a single
+ * RegExp match
+ */
+import multi from "multi-loader";
+
+/**
  * @description It moves all the require("style.css")s in entry chunks into
  * a separate single CSS file. So your styles are no longer inlined
  * into the JS bundle, but separate in a CSS bundle file (styles.css).
@@ -11,31 +17,48 @@ import webpack from "webpack";
  * the CSS bundle is loaded in parallel to the JS bundle.
  */
 import ExtractTextPlugin from "extract-text-webpack-plugin";
+
 /**
  * @description PostCSS plugin to parse CSS and add vendor prefixes
  * to CSS rules using values from Can I Use. It is recommended by Google
  * and used in Twitter and Taobao.
  */
 import autoprefixer from "autoprefixer";
+
+/**
+ * Get path from nodejs
+ */
 import path from "path";
+
+/**
+ * Get FS filesystem from nodejs
+ */
 import fs from "fs";
 
 import {
-  srcDir,
+  enableServiceWorker,
+  enableCommonStyles,
+  images,
+  isolateVendorScripts,
+} from "../settings";
+
+import {
   buildDir,
   buildPublicPath,
-  srcPublicDir, distPublicDir, distDir
+  distDir,
+  pagesDir,
+  srcDir,
+  srcPublicDir,
 } from "../directories";
 
-const pagesFolder = path.join(srcDir, "pages");
-const pages = fs.readdirSync(pagesFolder);
+const pages = fs.readdirSync(pagesDir);
 let entries = {};
 
 pages.forEach(page => {
-  const slugishName = page.replace(".js", "").replace(/['" \-!@#$%]/g, "_");
+  const slugishName = page.replace(".js", "").replace(/['" !@#$%]/g, "_");
   entries[`mod-${slugishName}`] = [
     "react-hot-loader/patch",
-    path.join(pagesFolder, page),
+    path.join(pagesDir, page),
     "webpack-hot-middleware/client?path=/__hmr_update&timeout=2000&overlay=true"
   ];
 });
@@ -53,18 +76,19 @@ const rules = [
       },
       {
         loader: "eslint-loader",
-        options: {
-          failOnWarning: false,
-          failOnError: false
-        }
       }
     ]
   },
+  
+  // Managing styles which are not present in resources folder
   {
-    test: /\.(sass|scss|css)$/, //Check for sass or scss file names,
-    exclude: [
-      path.join(srcDir, "resources"),
-    ],
+    test: /\.(sass|scss|css)$/, // Check for sass or scss file names,
+    ...(enableCommonStyles ? {
+      exclude: [
+        // We do not want to include sass|scss|css from resources with local ident
+        path.join(srcDir, "resources"),
+      ]
+    }: {}),
     use: [
       {
         loader: "style-loader",
@@ -94,40 +118,20 @@ const rules = [
       }
     ]
   },
-  {
-    test: /\.(eot|svg|ttf|woff|woff2)$/,
-    include: [
-      path.join(srcDir, "resources", "fonts"),
-    ],
-    use: "file-loader?outputPath=fonts/&name=[name].[ext]"
-  },
-  {
-    test: /\.svg$/i,
-    include: [
-      path.join(srcDir, "resources", "images"),
-    ],
-    use: [
-      "file-loader?outputPath=images/&name=[name].[ext]",
-    ]
-  },
-  {
-    test: /\.(jpe?g|png|gif)$/i,
-    include: [
-      path.join(srcDir, "resources", "images"),
-    ],
-    use: [
-      "file-loader?outputPath=images/&name=[name].webp",
-      "webp-loader?{quality: 80}",
-    ]
-  },
-  {
-    test: /\.(sass|scss|css)$/, //Check for sass or scss file names,
-    include: [
-      path.join(srcDir, "resources"),
-    ],
-    use: ExtractTextPlugin.extract({
-      fallback: "style-loader",
+  
+  // Managing styles present in resources folder, they do not need complex IdentName
+  // Just [local]
+  ...(enableCommonStyles ? [
+    {
+      test: /\.(sass|scss|css)$/, //Check for sass or scss file names,
+      include: [
+        path.join(srcDir, "resources"),
+      ],
       use: [
+        {
+          loader: "style-loader",
+          options: { sourceMap: true }
+        },
         {
           loader: "css-loader",
           options: {
@@ -150,8 +154,34 @@ const rules = [
             sourceMapContents: true,
           }
         }
-      ]
-    }),
+      ],
+    }
+  ]: {}),
+  
+  // Managing fonts
+  {
+    test: /\.(eot|ttf|woff|woff2)$/,
+    use: "file-loader?outputPath=fonts/&name=[name].[ext]"
+  },
+  
+  // Manage svg from overall application
+  {
+    test: /\.svg$/i,
+    use: [
+      "file-loader?outputPath=images/&name=[name].[ext]",
+    ]
+  },
+  
+  // Manage images
+  // Create webp copy of jpeg,png and gif images if useWebP is set to true
+  {
+    test: /\.(jpe?g|png|gif)$/i,
+    use: [
+      multi(...[
+        ...(images.useWebP ? ["file-loader?outputPath=images/&name=[name].[ext].webp!webp-loader?{quality: 80}"] : []),
+        "file-loader?outputPath=images/&name=[name].[ext]",
+      ])
+    ]
   },
 ];
 
@@ -171,7 +201,9 @@ const commonClientConfig = {
       path.join(srcDir, "client.js"),
       "webpack-hot-middleware/client?name=common-client&path=/__hmr_update&timeout=2000&overlay=true"
     ],
-    "common-style": path.join(srcDir, "resources", "css", "style.scss")
+    ...(enableCommonStyles ? {
+      "common-style": path.join(srcDir, "resources", "css", "style.scss")
+    }: {})
   }, entries),
   
   //These options determine how the different types of modules within
@@ -205,7 +237,7 @@ const commonClientConfig = {
     // the base of content, in our case its the "src/public" folder
     contentBase: srcPublicDir,
     
-    compress: true,
+    compress: false,
     
     // Show errors and warning on overlap
     overlay: {
@@ -214,9 +246,13 @@ const commonClientConfig = {
     },
   },
   
-  devtool: "source-map",
+  devtool: "eval-source-map",
   
   plugins: [
+  
+    new webpack.DefinePlugin({
+      "process.env.NODE_ENV": JSON.stringify(process.env.NODE_ENV || "development"),
+    }),
     
     // Hot module replacement for getting latest updates
     // thus no reload required
@@ -224,23 +260,27 @@ const commonClientConfig = {
     
     // Create common chunk of data
     // Break data in common so that we have minimum data to load
-    new webpack.optimize.CommonsChunkPlugin({
-      name: "commons-vendor",
-      filename: "common-0-vendor-[hash].js",
-      minChunks: function (module) {
-        
-        if(module.resource && (/^.*\.(css|scss|sass)$/).test(module.resource)) {
-          return false;
-        }
-        
-        // this assumes your vendor imports exist in the node_modules directory
-        return module.context &&
-          (
-            module.context.indexOf("node_modules") !== -1 ||
-            (module.resource && module.resource.indexOf("/src/client") !== -1)
-          );
-      },
-    }),
+    ...(
+      isolateVendorScripts ? [
+        new webpack.optimize.CommonsChunkPlugin({
+          name: "commons-vendor",
+          filename: "common-0-vendor-[hash].js",
+          minChunks: function (module) {
+            
+            if(module.resource && (/^.*\.(css|scss|sass)$/).test(module.resource)) {
+              return false;
+            }
+            
+            // this assumes your vendor imports exist in the node_modules directory
+            return module.context &&
+              (
+                module.context.indexOf("node_modules") !== -1 ||
+                (module.resource && module.resource.indexOf("/src/client") !== -1)
+              );
+          },
+        })
+      ]: []
+    ),
     
     // Extract the CSS so that it can be moved to CDN as desired
     // Also extracted CSS can be loaded parallel
@@ -264,6 +304,10 @@ const commonClientConfig = {
   ]
 };
 
+/**
+ * Service worker need to only worry about JavaScript, other thing should be
+ * available via cache or install activity
+ */
 const serviceWorkerConfig = {
   name: "service-worker",
   // The base directory, an absolute path, for resolving entry points
@@ -293,18 +337,18 @@ const serviceWorkerConfig = {
           },
           {
             loader: "eslint-loader",
-            options: {
-              failOnWarning: false,
-              failOnError: false
-            }
           }
         ]
       },
+  
       {
-        test: /\.(sass|scss)$/,
-        exclude: [
-          path.join(srcDir, "resources"),
-        ],
+        test: /\.(sass|scss|css)$/,
+        ...(enableCommonStyles ? {
+          exclude: [
+            // We do not want to include sass|scss|css from resources with local ident
+            path.join(srcDir, "resources"),
+          ]
+        }: {}),
         loader: ExtractTextPlugin.extract({
           fallback: "style-loader",
           use: [
@@ -335,6 +379,44 @@ const serviceWorkerConfig = {
           ]
         }),
       },
+  
+      // Managing styles present in resources folder, they do not need complex IdentName
+      // Just [local]
+      ...(enableCommonStyles? [
+        {
+          test: /\.(sass|scss|css)$/, //Check for sass or scss file names,
+          include: [
+            path.join(srcDir, "resources"),
+          ],
+          use: ExtractTextPlugin.extract({
+            fallback: "style-loader",
+            use: [
+              {
+                loader: "css-loader",
+                options: {
+                  modules: true,
+                  localIdentName: "[local]",
+                  sourceMap: true,
+                  minimize: false,
+                  importLoaders: 2
+                }
+              },
+              {
+                loader: "postcss-loader",
+                options: { sourceMap: true }
+              },
+              {
+                loader: "sass-loader",
+                options: {
+                  outputStyle: "expanded",
+                  sourceMap: true,
+                  sourceMapContents: true,
+                }
+              }
+            ]
+          }),
+        }
+      ]: {}),
     ],
   },
   output: {
@@ -356,28 +438,30 @@ const serviceWorkerConfig = {
     ],
   },
   
-  devtool: "source-map",
+  devtool: "eval-source-map",
   
   plugins: [
-  
+    new webpack.DefinePlugin({
+      "process.env.NODE_ENV": JSON.stringify(process.env.NODE_ENV || "development"),
+    }),
+    
     // Enable no errors plugin
     new webpack.NoEmitOnErrorsPlugin(),
-    
-    new ExtractTextPlugin("service-worker.min.css"),
-    
-    // Sass loader options for autoprefix
-    new webpack.LoaderOptionsPlugin({
-      options: {
-        context: "/",
-        sassLoader: {
-          includePaths: [srcDir]
-        },
-        postcss: function () {
-          return [autoprefixer];
-        }
-      }
-    })
+  
+    // Extract the CSS so that it can be moved to CDN as desired
+    // Also extracted CSS can be loaded parallel
+    new ExtractTextPlugin({
+      filename: "service-worker.min.css"
+    }),
   ]
 };
-// export default commonClientConfig;
-export default [commonClientConfig, serviceWorkerConfig];
+
+// Export the client config by default
+let exportConfig = commonClientConfig;
+
+// service worker is enabled for the application
+// export multiple config
+if (enableServiceWorker) {
+  exportConfig = [commonClientConfig, serviceWorkerConfig];
+}
+export default exportConfig;
