@@ -23,12 +23,16 @@ import ExtractTextPlugin from "extract-text-webpack-plugin";
 import autoprefixer from "autoprefixer";
 
 import {
+  isolateVendorScripts
+} from "../settings";
+
+import {
   srcDir,
   rootDir,
   srcPublicDir,
   distPublicDir,
   buildDir,
-  buildPublicPath
+  buildPublicPath,
 } from "../directories";
 
 import rules from "./prod.rules";
@@ -54,17 +58,16 @@ const pagesFolder = path.join(srcDir, "pages");
 const pages = fs.readdirSync(pagesFolder);
 let entries = {};
 pages.forEach(page => {
-  const slugishName = page.replace(".js", "").replace(/['" \-!@#$%]/g, "_");
+  const slugishName = page.replace(".js", "").replace(/['" !@#$%]/g, "_");
   entries[`mod-${slugishName}`] = path.join(pagesFolder, page);
 });
 
-
 export default {
-
+  
   // The base directory, an absolute path, for resolving entry points
   // and loaders from configuration. Lets keep it to /src
   context: srcDir,
-
+  
   // The point or points to enter the application. At this point the
   // application starts executing. If an array is passed all items will
   // be executed.
@@ -75,34 +78,34 @@ export default {
       path.join(srcDir, "resources", "css", "style.scss")
     ]
   }, entries),
-
+  
   //These options determine how the different types of modules within
   // a project will be treated.
   module: {
     rules: rules({}),
   },
-
+  
   resolve: {
     modules: [
       "node_modules",
       srcDir
     ],
   },
-
+  
   output: {
-
+    
     // Output everything in dist folder
     path: buildDir,
-
+    
     // The file name to output
     filename: "[name].[chunkhash].bundle.js",
-
+    
     // public path is assets path
     publicPath: buildPublicPath,
   },
-
+  
   devtool: false,
-
+  
   plugins: [
     // While building remove the dist dir
     new CleanWebpackPlugin(["dist"], {
@@ -111,49 +114,53 @@ export default {
     }),
     // Assets plugin to generate
     AssetsPluginInstance,
-
+    
     // Uglify the output so that we have the most optimized code
     new UglifyJSPlugin({
       compress: true,
       comments: false,
       sourceMap: false,
     }),
-
+    
     // Create common chunk of data
     // Break data in common so that we have minimum data to load
-    new webpack.optimize.CommonsChunkPlugin({
-      name: "commons-vendor",
-      filename: "common-vendor-[chunkhash].js",
-      minChunks: function (module) {
-        /** Ignore css files **/
-        if(module.resource && (/^.*\.(css|scss|sass)$/).test(module.resource)) {
-          return false;
-        }
-        // this assumes your vendor imports exist in the node_modules directory
-        return module.context &&
-          (
-            module.context.indexOf("node_modules") !== -1 ||
-            (module.resource && module.resource.indexOf("/src/client") !== -1)
-          );
-      },
-    }),
-
+    ...(
+      isolateVendorScripts ? [
+        new webpack.optimize.CommonsChunkPlugin({
+          name: "commons-vendor",
+          filename: "common-vendor-[chunkhash].js",
+          minChunks: function (module) {
+            /** Ignore css files **/
+            if(module.resource && (/^.*\.(css|scss|sass)$/).test(module.resource)) {
+              return false;
+            }
+            // this assumes your vendor imports exist in the node_modules directory
+            return module.context &&
+              (
+                module.context.indexOf("node_modules") !== -1 ||
+                (module.resource && module.resource.indexOf("/src/client") !== -1)
+              );
+          },
+        }),
+      ]: []
+    ),
+    
     new CopyWebpackPlugin([{
       from: srcPublicDir,
       to: distPublicDir,
     }]),
-
+    
     new webpack.DefinePlugin({
-      "process.env.NODE_ENV": JSON.stringify("production")
+      "process.env.NODE_ENV": JSON.stringify(process.env.NODE_ENV || "production")
     }),
-
+    
     // Enable no errors plugin
     new webpack.NoEmitOnErrorsPlugin(),
-
+    
     // Extract the CSS so that it can be moved to CDN as desired
     // Also extracted CSS can be loaded parallel
     new ExtractTextPlugin("[name].[chunkhash].min.css"),
-
+    
     // Sass loader options for autoprefix
     new webpack.LoaderOptionsPlugin({
       options: {
