@@ -1,6 +1,8 @@
 import path from "path";
 import webpack from "webpack";
 import UglifyJSPlugin from "uglifyjs-webpack-plugin";
+import WebpackDeleteAfterEmit from "webpack-delete-after-emit";
+
 
 /**
  * @description It moves all the require("style.css")s in entry chunks into
@@ -18,12 +20,10 @@ import ExtractTextPlugin from "extract-text-webpack-plugin";
  */
 import autoprefixer from "autoprefixer";
 
-import {
-  srcDir,
-  distDir, distPublicDir,
-} from "../directories";
+import { srcDir, distDir } from "../directories";
 
 import rules from "./prod.rules";
+import {enableServiceWorker} from "../settings";
 
 export default [{
   
@@ -31,16 +31,16 @@ export default [{
   // The base directory, an absolute path, for resolving entry points
   // and loaders from configuration. Lets keep it to /src
   context: srcDir,
-
+  
   // The point or points to enter the application. At this point the
   // application starts executing. If an array is passed all items will
   // be executed.
   entry: [
     "babel-polyfill",
     // Initial entry point
-    path.join(srcDir, "start-server.js"),
+    path.join(srcDir, "core/start-server.js"),
   ],
-
+  
   //These options determine how the different types of modules within
   // a project will be treated.
   module: {
@@ -53,26 +53,26 @@ export default [{
     ],
   },
   output: {
-
+    
     // Output everything in dist folder
     path: distDir,
-
+    
     // The file name to output
     filename: "server.js",
-
+    
     // public path is assets path
     publicPath: "/",
   },
-
+  
   node: {
     __filename: true,
     __dirname: true
   },
   target: "node",
   devtool: false,
-
+  
   plugins: [
-
+    
     // Uglify the output so that we have the most optimized code
     new UglifyJSPlugin({
       compress: true,
@@ -80,11 +80,11 @@ export default [{
       sourceMap: false,
     }),
     new webpack.DefinePlugin({
-      "process.env.NODE_ENV": JSON.stringify("production"),
+      "process.env.NODE_ENV": JSON.stringify(process.env.NODE_ENV || "production"),
     }),
     // Enable no errors plugin
     new webpack.NoEmitOnErrorsPlugin(),
-
+    
     // Extract the CSS so that it can be moved to CDN as desired
     // Also extracted CSS can be loaded parallel
     new ExtractTextPlugin("server.min.css"),
@@ -99,77 +99,96 @@ export default [{
           return [autoprefixer];
         }
       }
+    }),
+    // We are extracting server.min.css so that we do not have any window code in server.js
+    // but we still need the css class names that are generated. Thus we remove the server.min.css
+    // after the build process
+    new WebpackDeleteAfterEmit({
+      globs: ["server.min.css"]
     })
   ],
 },
-{
-  name: "service-worker",
+...(enableServiceWorker ? [
+  {
+    name: "service-worker",
+  
+    // The base directory, an absolute path, for resolving entry points
+    // and loaders from configuration. Lets keep it to /src
+    context: srcDir,
+  
+    // The point or points to enter the application. At this point the
+    // application starts executing. If an array is passed all items will
+    // be executed.
+    entry: {
+      "service-worker" : [
+        "babel-polyfill",
+        // Initial entry point
+        path.join(srcDir, "service-worker.js"),
+      ]
+    },
+  
+    // These options determine how the different types of modules within
+    // a project will be treated.
+    module: {
+      rules: rules({}),
+    },
+    resolve: {
+      modules: [
+        "node_modules",
+        srcDir
+      ],
+    },
+    output: {
     
-  // The base directory, an absolute path, for resolving entry points
-  // and loaders from configuration. Lets keep it to /src
-  context: srcDir,
+      // Output everything in dist folder
+      path: distDir,
     
-  // The point or points to enter the application. At this point the
-  // application starts executing. If an array is passed all items will
-  // be executed.
-  entry: {
-    "service-worker" : [
-      "babel-polyfill",
-      // Initial entry point
-      path.join(srcDir, "service-worker.js"),
-    ]
-  },
+      // The file name to output
+      filename: "[name].js",
     
-  // These options determine how the different types of modules within
-  // a project will be treated.
-  module: {
-    rules: rules({}),
-  },
-  resolve: {
-    modules: [
-      "node_modules",
-      srcDir
-    ],
-  },
-  output: {
-      
-    // Output everything in dist folder
-    path: distDir,
-      
-    // The file name to output
-    filename: "[name].js",
-      
-    // public path is assets path
-    publicPath: "/",
-  },
-  target: "web",
-  devtool: false,
+      // public path is assets path
+      publicPath: "/",
+    },
+    target: "web",
+    devtool: false,
+  
+    plugins: [
     
-  plugins: [
-      
-    // Uglify the output so that we have the most optimized code
-    new UglifyJSPlugin({
-      compress: true,
-      comments: false,
-      sourceMap: false,
-    }),
-    new webpack.DefinePlugin({
-      "process.env.NODE_ENV": JSON.stringify("production"),
-    }),
-    // Enable no errors plugin
-    new webpack.NoEmitOnErrorsPlugin(),
+      // Uglify the output so that we have the most optimized code
+      new UglifyJSPlugin({
+        compress: true,
+        comments: false,
+        sourceMap: false,
+      }),
+      new webpack.DefinePlugin({
+        "process.env.NODE_ENV": JSON.stringify(process.env.NODE_ENV || "production"),
+      }),
+      // Enable no errors plugin
+      new webpack.NoEmitOnErrorsPlugin(),
     
-    // Sass loader options for autoprefix
-    new webpack.LoaderOptionsPlugin({
-      options: {
-        context: "/",
-        sassLoader: {
-          includePaths: [srcDir]
-        },
-        postcss: function () {
-          return [autoprefixer];
+      // Extract the CSS so that it can be moved to CDN as desired
+      // Also extracted CSS can be loaded parallel
+      new ExtractTextPlugin("service-worker.min.css"),
+    
+      // Sass loader options for autoprefix
+      new webpack.LoaderOptionsPlugin({
+        options: {
+          context: "/",
+          sassLoader: {
+            includePaths: [srcDir]
+          },
+          postcss: function () {
+            return [autoprefixer];
+          }
         }
-      }
-    })
-  ],
-}];
+      }),
+  
+      // We are extracting server.min.css so that we do not have any window code in service-worker.js
+      // but we still need the css class names that are generated. Thus we remove the server.min.css
+      // after the build process
+      new WebpackDeleteAfterEmit({
+        globs: ["service-worker.min.css"]
+      })
+    ],
+  }
+] : [])];
